Increased trigger delay range.  It has to be *greater* than a sampling quantity, not less.
Added a call to setBaselineSubtraction, which is not initialized by the configuration.
Without doing this, the IPIMB record is stuck at raw 0xffff.

Fixed path to event screen in edm-ioc-tst-05.cmd.

Added evrEventAll.db to the database, otherwise we do not get the records that process
the timestamps by EvrTimeEvent.


git-svn-id: file:///afs/slac/g/pcds/vol2/svn/pcds/epics/trunk/modules/ipimb/current@7512 ececa60f-9f63-0410-97c7-ea05888eb260